splitup code into multiple files
